added test_cov and badges
