+++ .github/workflows/ci-cd.yml
@@ -88,17 +88,13 @@ jobs:
 
     - name: Run unit tests
       run: |
-        pytest cmaclp/tests/cmaclp_test.py -vvv --disable-pytest-warnings \
+        pytest cmaclp/tests/SVM_prediction_test.py -vvv --disable-pytest-warnings \
         --cov=./ --cov-report=xml
 
     - name: Upload code coverage ☂️
       uses: paambaati/codeclimate-action@v4.0.0
       env:
         CC_TEST_REPORTER_ID: ${{ secrets.CC_TEST_REPORTER_ID }}
-        
-    # Running end-to-end tests CLI functions
-    - name: Running end-to-end testing    
-      run: pytest cmaclp/tests/cmaclp_test_end_to_end.py --cov
 
   cd:
     # Only run this job if the "ci" job passes
