+++ README.md
@@ -1,4 +1,6 @@
 [![CI/CD](https://github.com/Arts-of-coding/cmaclp/actions/workflows/ci-cd.yml/badge.svg)](https://github.com/Arts-of-coding/cmaclp/actions/workflows/ci-cd.yml)
+[![Maintainability](https://api.codeclimate.com/v1/badges/598ba117b586183c46a8/maintainability)](https://codeclimate.com/github/Arts-of-coding/cmaclp/maintainability)
+[![Test Coverage](https://api.codeclimate.com/v1/badges/598ba117b586183c46a8/test_coverage)](https://codeclimate.com/github/Arts-of-coding/cmaclp/test_coverage)
 # cma
 This repository defines a command-line tool to predict (clp) datasets according to a cell meta-atlas (cma). At the present time only the meta-atlas for the cornea has been implemented.
