+++ cmaclp/SVM_prediction.py
@@ -759,4 +759,3 @@ if __name__ == '__performpars__':
 
 if __name__ == '__importpars__':
     importpars()
-
