+++ cmaclp/__init__.py
@@ -1,0 +1,8 @@
+__version__ = "0.1.1"
+
+from .SVM_prediction import SVM_prediction
+from .SVM_prediction import SVM_import
+from .SVM_prediction import SVM_pseudobulk
+from .SVM_prediction import predpars
+from .SVM_prediction import importpars
+from .SVM_prediction import performpars
