+++ cmaclp/tests/SVM_prediction_test.py
@@ -1,4 +1,5 @@
 # Import modules
+# Import modules
 import argparse
 import os
 import numpy as np
@@ -6,7 +7,7 @@ import pandas as pd
 import scanpy as sc
 import time as tm
 import seaborn as sns
-import cmaclp
+#import cmaclp
 from sklearn.svm import LinearSVC
 from sklearn.calibration import CalibratedClassifierCV
 from sklearn.metrics import confusion_matrix
@@ -23,6 +24,11 @@ from scipy.stats import spearmanr
 from cmaclp.SVM_prediction import SVM_prediction
 from cmaclp.SVM_prediction import SVM_import
 from cmaclp.SVM_prediction import SVM_pseudobulk
+
+# Perhaps add?
+from cmaclp.SVM_prediction import predpars
+from cmaclp.SVM_prediction import importpars
+from cmaclp.SVM_prediction import performpars
 
 # OLD
 #def obs_key_wise_subsampling(adata, obs_key, N):
@@ -60,23 +66,21 @@ outdir_unit = "test_output_unit/"
 
 ### UNIT TESTS
 
-def test_unit_SVM_prediction():
-
+def test_SVM_prediction():
     SVM_prediction(reference_H5AD=reference,query_H5AD=query,
       LabelsPathTrain=labels,OutputDir=outdir_unit)
 
     assert os.path.exists(f'{outdir_unit}SVM_Pred_Labels.csv') == 1
 
 
-def test_unit_SVMrej_prediction():
-
+def test_SVMrej_prediction():
     SVM_prediction(reference_H5AD=reference,query_H5AD=query,
       LabelsPathTrain=labels,OutputDir=outdir_unit,rejected=True)
 
     assert os.path.exists(f'{outdir_unit}SVMrej_Pred_Labels.csv') == 1
 
     
-def test_unit_SVM_import():
+def test_SVM_import():
 
     SVM_import(query_H5AD=query,OutputDir=outdir_unit,SVM_type="SVM",replicates="time_point",meta_atlas=True)
 
@@ -84,7 +88,7 @@ def test_unit_SVM_import():
     assert os.path.exists(f'SVM_predicted.h5ad') == 1
 
 
-def test_unit_SVM_import_plots():
+def test_SVM_plot_import():
 
     SVM_import(query_H5AD=query,OutputDir=outdir_unit,
       SVM_type="SVM",replicates="time_point",show_bar=True)
@@ -92,10 +96,68 @@ def test_unit_SVM_import_plots():
     assert os.path.exists(f'figures/SVM_predicted_bar.pdf') == 1
 
       
-def test_unit_SVM_pseudobulk():
+def test_SVM_pseudobulk():
 
     SVM_pseudobulk(condition_1=reference, condition_1_batch="donors", 
       condition_2="SVM_predicted.h5ad", condition_2_batch="batch", Labels_1=labels)
     assert os.path.exists("pseudobulk_output/full_batch_samples.tsv") == 1
     assert os.path.exists("pseudobulk_output/merged_batch_samples.tsv") == 1
+
+### END-TO-END
+outdir = "test_output_end_to_end/"
+
+def test_CLI_SVM_prediction():
+
+    command_to_be_executed = ['SVM_prediction',
+                              '--reference_H5AD', str(reference),
+                              '--query_H5AD', str(query),
+                              '--LabelsPathTrain', str(labels),
+                              '--OutputDir', str(outdir)]
+
+    subprocess.run(command_to_be_executed, shell=False, timeout=None,
+                   text=True)
+    assert os.path.exists(f'{outdir}SVM_Pred_Labels.csv') == 1
+
+
+def test_CLI_SVMrej_prediction():
+
+    command_to_be_executed = ['SVM_prediction',
+                              '--reference_H5AD', str(reference),
+                              '--query_H5AD', str(query),
+                              '--LabelsPathTrain', str(labels),
+                              '--OutputDir', str(outdir),
+                              '--rejected']
+
+    subprocess.run(command_to_be_executed, shell=False, timeout=None,
+                   text=True)
+    assert os.path.exists(f'{outdir}SVMrej_Pred_Labels.csv') == 1
+
+    
+def test_CLI_SVM_import():
+
+    command_to_be_executed = ['SVM_import',
+                              '--query_H5AD', str(query),
+                              '--OutputDir', str(outdir),
+                              '--SVM_type', str("SVM"),
+                              '--meta-atlas']
+
+    subprocess.run(command_to_be_executed, shell=False, timeout=None,
+                   text=True)
+    assert os.path.exists(f'figures/Density_prediction_scores.pdf') == 1
+    assert os.path.exists(f'SVM_predicted.h5ad') == 1
+
+
+def test_CLI_SVM_import_plots():
+
+    command_to_be_executed = ['SVM_import',
+                              '--query_H5AD', str(query),
+                              '--OutputDir', str(outdir),
+                              '--SVM_type', str("SVM"),
+                              '--replicates', str("time_point"),
+                              '--show-bar']
+
+    subprocess.run(command_to_be_executed, shell=False, timeout=None,
+                   text=True)
+    assert os.path.exists(f'figures/SVM_predicted_bar.pdf') == 1
+      #assert os.path.exists(f'figures/umap_SVM_predicted.pdf') == 1
